+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0+4] - 2024-03-19
+
+### Added
+- Symptom intensity tracking (mild, moderate, strong)
+- Night time option for symptoms
+
+### Changed
+- Existing symptoms default to moderate intensity
+
 ## [1.0.0+3] - 2024-03-19
 
 ### Fixed